requirements.txt
# Data and API
requests==2.31.0
pandas==2.1.3
numpy==1.24.3

# Technical Analysis
ta==0.11.0
scipy==1.11.4

# Visualization and Dashboard
plotly==5.18.0
dash==2.14.1
dash-bootstrap-components==1.5.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Data storage
sqlalchemy==2.0.23